ChatBot/requirements.txt
urllib3==1.24.3
openai==0.27.6
python-dotenv==1.0.0
langchain==0.0.176
Django==4.2.1
aiohttp==3.8.3
Brotli==1.0.9
urllib3==1.24.3
mysqlclient==1.4.3
flask
gunicorn
gevent
edge-tts
djangorestframework
django-cors-headers
channels
daphne
schedule
drf-yasg

